# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StatusResponse {
  status: string;
  db: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  Scheduled: 'Scheduled',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
  Rescheduled: 'Rescheduled',
} as const;

export interface Appointment {
  id: string;
  clientName: string;
  service: string;
  date: string;
  time: string;
  location?: string;
  caregiverId?: string;
  status: AppointmentStatus;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export type AppointmentCreateStatus = typeof AppointmentCreateStatus[keyof typeof AppointmentCreateStatus];


export const AppointmentCreateStatus = {
  Scheduled: 'Scheduled',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
  Rescheduled: 'Rescheduled',
} as const;

export interface AppointmentCreate {
  clientName: string;
  service: string;
  date: string;
  time: string;
  location?: string;
  caregiverId?: string;
  status?: AppointmentCreateStatus;
  notes?: string;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  Scheduled: 'Scheduled',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
  Rescheduled: 'Rescheduled',
} as const;

export interface AppointmentUpdate {
  clientName?: string;
  service?: string;
  date?: string;
  time?: string;
  location?: string;
  caregiverId?: string;
  status?: AppointmentUpdateStatus;
  notes?: string;
}

export interface Message {
  id: string;
  senderId: string;
  recipientId: string;
  subject?: string;
  text: string;
  sentAt: string;
  read: boolean;
}

export interface MessageCreate {
  senderId: string;
  recipientId: string;
  subject?: string;
  text: string;
}

export type NotificationCategory = typeof NotificationCategory[keyof typeof NotificationCategory];


export const NotificationCategory = {
  alert: 'alert',
  reminder: 'reminder',
  update: 'update',
  info: 'info',
} as const;

export interface Notification {
  id: string;
  title: string;
  body: string;
  category: NotificationCategory;
  createdAt: string;
  read: boolean;
}

export type NotificationCreateCategory = typeof NotificationCreateCategory[keyof typeof NotificationCreateCategory];


export const NotificationCreateCategory = {
  alert: 'alert',
  reminder: 'reminder',
  update: 'update',
  info: 'info',
} as const;

export interface NotificationCreate {
  title: string;
  body: string;
  category: NotificationCreateCategory;
}

export interface CareTeamMember {
  id: string;
  name: string;
  role: string;
  status: string;
  specialties: string[];
  rating: number;
  bio?: string;
  avatarUrl?: string;
  isPrimary: boolean;
  updatedAt: string;
}

export interface CareTeamMemberCreate {
  name: string;
  role: string;
  status: string;
  specialties: string[];
  rating: number;
  bio?: string;
  avatarUrl?: string;
  isPrimary: boolean;
}

export interface CareTeamMemberUpdate {
  name?: string;
  role?: string;
  status?: string;
  specialties?: string[];
  rating?: number;
  bio?: string;
  avatarUrl?: string;
  isPrimary?: boolean;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  Paid: 'Paid',
  Pending: 'Pending',
  Overdue: 'Overdue',
} as const;

export interface Invoice {
  id: string;
  invoiceId: string;
  date: string;
  dueDate: string;
  description: string;
  amount: number;
  status: InvoiceStatus;
  clientName?: string;
  createdAt: string;
}

export type InvoiceCreateStatus = typeof InvoiceCreateStatus[keyof typeof InvoiceCreateStatus];


export const InvoiceCreateStatus = {
  Paid: 'Paid',
  Pending: 'Pending',
  Overdue: 'Overdue',
} as const;

export interface InvoiceCreate {
  invoiceId: string;
  date: string;
  dueDate: string;
  description: string;
  amount: number;
  status: InvoiceCreateStatus;
  clientName?: string;
}

export interface HealthReport {
  id: string;
  type: string;
  value: string;
  status: string;
  date: string;
  notes?: string;
  createdAt: string;
}

export interface HealthReportCreate {
  type: string;
  value: string;
  status: string;
  date: string;
  notes?: string;
}

export interface ErrorResponse {
  error: string;
}