# packages/api-zod/src/generated/types/appointmentCreate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentCreateStatus } from './appointmentCreateStatus';

export interface AppointmentCreate {
  clientName: string;
  service: string;
  date: Date;
  time: string;
  location?: string;
  caregiverId?: string;
  status?: AppointmentCreateStatus;
  notes?: string;
}