# packages/api-zod/src/generated/types/careTeamMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareTeamMember {
  id: string;
  name: string;
  role: string;
  status: string;
  specialties: string[];
  rating: number;
  bio?: string;
  avatarUrl?: string;
  isPrimary: boolean;
  updatedAt: Date;
}